fix: test that (car head) is specified before computing inherit
Some themes may be defined such that ='unspecified= is the car of head resulting
to ~(face-attribute 'unspecified :inherit)~ being evaluated which raises the
error =Invalid face: unspecified= since unspecified is an invalid face.

Since we don't control how themes are defined, we should at least control that
we don't end up in an error state in the htmlize logic because of an
"incomplete" theme.

I do not fully understand how =htmlize-face-size= works BTW since I find the
function difficult to reason about. I've ended up retrofitting it with a bunch
of =message= calls (yes, I'm "debugging by printf" like a shameless noob 🙊) and
am really struggling with some of the state changes that are taking place.
Changes to the =face-list=, =head= and =tail= variables are not always
explicitly expressed and in some changes happen through side-effects of =setcdr=
and =pop= which complicate reasoning about this logic.

Looking at the implementation, I kept wondering how ~(htmlize-face-size face)~
relates to calling ~(face-attribute face :height nil t)~ which in my setup
returns the same value. Perhaps this function exists for historic reasons, in
which case we may be able to refactor this out for a simpler
implementation (e.g.: just using =face-attribute=). Looking at the git blame, it
seems like the htmlize-face-size implementation
https://github.com/hniksic/emacs-htmlize/blame/master/htmlize.el#L1029-L1050
succeeded the face-attribute implementation
https://github.com/emacs-mirror/emacs/blame/master/lisp/faces.el#L450-L487 which
would lead me to believe that "historic reasons" may not quite apply here but
asking for context won't hurt.